Modify schedule functionality added

+++ scheduler/views.py
@@ -12,27 +12,33 @@ from manager.models import *
 from django_celery_beat.models import CrontabSchedule, PeriodicTask
 import json
 import os
-from bacchus import settings
-# Create your views here.
+from celerytools import CeleryTools
 
-@login_required(login_url="/login/")
-def restart_beat():
-    return os.system('cd '+settings.SCRIPTS_DIR+';./'+settings.BEAT_RESTART_SCRIPT)
+# Create your views here.
 
 @login_required(login_url="/login/")
 def list_sched(request):
     if request.method == "POST":
-        sched_id = request.POST.get('sched_id')
-        sched = Schedule.objects.get(id=sched_id)
-        # delete celery schedule entry
-        
-        # delete db schedule entries
-        backupscheds = BackupSchedule.objects.all().filter(schedule=sched)
-        for backupsched in backupscheds:
-            backupsched.delete()
-        
-        sched.delete()
-        restart_beat()
+        if request.POST.get('delete_sched') == "delete":
+
+            sched_id = request.POST.get('sched_id')
+            sched = Schedule.objects.get(id=sched_id)
+            # delete celery schedule entries
+            celery_task = PeriodicTask.objects.get(id=sched.celery_task_id)
+            celery_task.delete()
+            
+            celery_sched = CrontabSchedule.objects.get(id=sched.celery_sched_id)
+            celery_sched.delete()
+            
+            # delete db schedule entries
+            backupscheds = BackupSchedule.objects.all().filter(schedule=sched)
+            for backupsched in backupscheds:
+                backupsched.delete()
+            
+            sched.delete()
+            CeleryTools.restart_beat()
+        else:
+            return redirect('/editsched/'+request.POST.get('sched_id'))
         
     scheds = Schedule.objects.all()
     
@@ -41,6 +47,7 @@ def list_sched(request):
 @login_required(login_url="/login/")
 def define_sched(request):
     if request.method == "POST":
+        
         sched_name = request.POST.get('sched_name')
         sched_min = request.POST.get('sched_min')
         sched_hour = request.POST.get('sched_hour')
@@ -51,19 +58,24 @@ def define_sched(request):
         sched.save()
         celery_sched = CrontabSchedule.objects.create(minute=sched_min,hour=sched_hour,day_of_week=sched_week,day_of_month=sched_month,month_of_year=sched_year)
         sched_args = json.dumps([sched.id])
+        sched.celery_sched_id = celery_sched.id
+        sched.save()
         
-        PeriodicTask.objects.create(crontab=celery_sched,name=sched_name,task='scheduler.tasks.run_schedule',args=sched_args)
+        celery_task = PeriodicTask.objects.create(crontab=celery_sched,name=sched_name,task='scheduler.tasks.run_schedule',args=sched_args)
+        sched.celery_task_id = celery_task.id
+        sched.save()
         
         # refresh celery_beat here to reflect the changes
-        restart_beat()
+        CeleryTools.restart_beat()
         
         vmlist = request.POST.getlist('schedvmlist')
         for vmid in vmlist:
             vm = VM.objects.get(vmid=vmid)
             backupsched = BackupSchedule(vm=vm,schedule=sched)
-            backupsched.save()            
-        scheds = Schedule.objects.all()
-        return render(request,'schedules.html',{'scheds': scheds })
+            backupsched.save()      
+        
+        return redirect('/schedules/')
+
     else:
         vms = VM.objects.all()
         return render(request,'definesched.html',{'vms': vms })
@@ -71,7 +83,7 @@ def define_sched(request):
 
 @login_required(login_url="/login/")
 def list_maint(request):
-    list_maint = 0
+    list_maint = 0  
     return render(request,'listmaint.html',{'list_maint': list_maint })
 
 @login_required(login_url="/login/")
@@ -81,5 +93,23 @@ def delete_sched(request):
         
     scheds = Schedule.objects.all()
     return render(request,'schedules.html',{'scheds': scheds })
+
+
+@login_required(login_url="/login/")
+def edit_sched(request, sched_id):
+       
+    sched = Schedule.objects.get(id=sched_id)
+    minutes = range(0,60)
+    minutes.insert(0,'*')
+    hours = range(0,24)
+    hours.insert(0,'*')
+    days = range(1,32)
+    days.insert(0,'*')
+    months = range(1,13)
+    months.insert(0,'*')
+    
+    return render(request,'editsched.html',{'sched': sched,'minutes': minutes, 'hours': hours, 'days':days, 'months':months })
+
+
